requirements.txt
# beefirst - Production Dependencies
# Trust State Machine Registration API

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg[binary]>=3.1.0
bcrypt>=4.1.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-dotenv>=1.0.0
